Update SwiftMatrixSDK to v0.11.6
The SwiftMatrixSDK developers have documented all the breaking changes between v0.10 and v0.11 of SwiftMatrixSDK in the release notes for v0.11.0. Those notes are available on GitHub at the following address <https://github.com/matrix-org/matrix-ios-sdk/releases/tag/v0.11.0>

Only two of them appear to impact Seaglass, so I'm summarizing those here:

1) MXRoom: liveTimeline and state accesses are now asynchronous.

Most of the changes to support this involved wrapping the code that accessed the async property in a closure. There were a few places that required further rework, like adding state as an initialize to an object or a parameter to a method so they could be available when the method returned.

2) MXRoomState: Create an MXRoomMembers property. All members getter methods has been to the new class.

The MXRoomState changes to the member class required adding an extra `.members` in a few places to get access to the `members` array that used to be directly on `MXRoomState` but is not in the `MXRoomMembers` property on `MXRoomState`.

I forked swift-matrix-sdk to cherry-pick a change that disabled AVAudioSession. Updating xCode caused this to break. Its fixed in newer versions of swift-matrix-sdk, once that library is up-to-date the Podfile can be changed to use mainline swift-matrix-sdk again.